Use different severity for client errors
The `ERROR` severity in logs is used for things that developers MUST
track and fix when possible. However, the 4xx status codes are related
to client mistakes - and the API developers are usually unable to solve.

By using `WARNING` for 4xx we allow users of this middleware to have the
opportunity to fine-tune their production logging and potentially ignore
client errors.